Refactor StorageService
Now uses different storage method. Appends a specified "store" name to each object id and uses this as a key, instead of storing all objects of a certain type under a single key.  This allows for a single object to be retrieved/saved without the need to retrieve all objects, but still allows to retrieve all objects of a certain type (just look at the start of every key).